+++ app/erp/sale_document/service_payment_invoice.py
@@ -1,7 +1,7 @@
 import logging
 import uuid
 
-from common.enums.sale_document_enum import SaleDocumentPaymentStatus
+from common.enums.sale_document_enum import SaleDocumentPaymentStatus, SaleDocumentStatus
 from django.apps import apps
 from django.db import transaction
 from django.http import Http404
@@ -16,8 +16,12 @@ def verify_transaction_public_ids(public_id):
     return False
 
 
+def verify_status_is_confirmed(sale_document):
+    return sale_document.status == SaleDocumentStatus.CONFIRMED.name
+
+
 def handle_invoice_payment(sale_document, transaction_public_ids):
-    if verify_transaction_public_ids(transaction_public_ids):
+    if verify_transaction_public_ids(transaction_public_ids) and verify_status_is_confirmed(sale_document):
         try:
             delete_old_paid_transaction_sale_document(sale_document)
             update_amount_paid_document(sale_document, 0, SaleDocumentPaymentStatus.NOT_PAID.name)
@@ -27,6 +31,9 @@ def handle_invoice_payment(sale_document, transaction_public_ids):
             logger.error(f"An error occurred while saving payment invoice: {ex}", exc_info=True,
                          extra={'object_id': object.id})
             raise Http404({"error": _("Erreur lors l'enregistrement du paiement")})
+        
+    if not verify_status_is_confirmed(sale_document):
+        delete_old_paid_transaction_sale_document(sale_document)
 
 
 def delete_old_paid_transaction_sale_document(sale_document):
